requirements.txt
fastapi==0.115.11
fpdf2==2.8.2
Jinja2==3.1.5
keybert==0.9.0
minio==7.2.15
numpy==2.2.4
pdf2image==1.17.0
pdfminer.six==20240706
python-dotenv==1.0.1
python-multipart==0.0.20
pyvis==0.3.2
scikit-learn==1.6.1
scipy==1.15.2
sentence-transformers==3.4.1
starlette==0.45.3
structlog==25.2.0
transformers==4.49.0
urllib3==2.3.0
uvicorn==0.34.0
pymorphy3